Add TRANS all info query
This query extracts all the info available for TRANS classes, even
though not all of this info is imported.

TRANS lacks all the elements not present in SYMP and additionally
does not currently have alternateID/'term replaced by' annotations,
cross-references, or comments.

Deprecated entities exist but are excluded from import.